custom_dragon_model
I made a custom dragon model to use for our rpg that I made through sculptris. This is if you want to use a better dragon model than the one already in the scene and you cannot find good free models online. I can also teach people how to use sculptris as it is a very free and easy tool.