Orc Movement
Made an orc movement script for the orc. NOTE: orc model must be rigged and have animations for this to work. add the animation names in the public fields of the script. also add the healthbar as well to the public field. Use the orc berserker model as it is already rigged. I recommend setting the model to legacy and then use animation component instead of animator.